+++ src/main/java/io/anserini/reproduce/RunMsMarco.java
@@ -1,5 +1,19 @@
-package io.anserini.tcr;
-
+package io.anserini.reproduce;
+/*
+ * Anserini: A Lucene toolkit for reproducible information retrieval research
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
@@ -15,7 +29,7 @@ public class RunMsMarco {
 
   public static void main(String[] args) throws Exception {
     final ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
-    Config config = mapper.readValue(new FileInputStream("src/main/java/io/anserini/tcr/msmarco-v1-passage.yaml"), Config.class);
+    Config config = mapper.readValue(new FileInputStream("src/main/java/io/anserini/reproduce/msmarco-v1-passage.yaml"), Config.class);
     TrecEvalMetricDefinitions metricDefinitions = new TrecEvalMetricDefinitions();
 
     for (Condition condition : config.conditions) {
